17543: no softfail on IOErrors in check_media_file
We shouldn't generally be getting IOErrors in read_metadata, but when we
do it's generally because another process is messing with our files.
Since it's outside our control, changing behavior in that case from a
soft-fail to a warning, and recovering otherwise as before.

This also affects handling of the _allow_nonexistent_paths property,
which is only set in unittests: it was affecting only files with known
non-media extensions, which isn't very useful. After this change, the
property does nothing. This doesn't seem to affect any unittests, and
can't affect anything outside of unittests.