Added handling for spawning subprocesses.
Added generic code to spawn a subprocess, send/receive messages for it, and
restart it if it crashes.  This is in preparation of moving feedparser
processing to a subprocess (#17216).

Moved autodler initialization code to the start_download() method.  It makes
more sense there and makes it easier to import modules in the subprocess.

Refactored config code used for the downloader and used it for subprocesses as
well.