Run feedparser in another process.
Replaced old code that runs feedparser in a thread with code that runs
feedparser in a separate process.  Because of the python GIL, this results in
better UI responsiveness.

Added flag in feed.py to skip the worker process and just run feedparser
inline.  This makes things simpler and faster in the unittests.

Removed really old code that changes how we pickle FeedParserDicts.  It was
written when our database was pickled objects.  Now it's just getting in the
way.

Removed old feedparser queue code.  workerprocess handles things much better.