Fix for next_free_filename()
The code that split the extension failed pretty bad if the file didn't have a
"." in it, but a directory leading up to it did.  For example, if we were
trying to generate names for "/home/ben/.miro/icon-cache/foo", then we would
look at these candidates:

"/home/ben/.miro/icon-cache/foo"
"/home/ben/.1.miro/icon-cache/foo"
"/home/ben/.2.miro/icon-cache/foo"

And never succeed because we were trying to create files in non-existent
directories.