Couple of fixes for #17731
Don't wrap IconCacheUpdater.request_update() with as_idle.  As far as I can
tell that method is only called from the backend thread anyway, so there's no
need for an idle callback.  It's called once per item when we create them, and
it's not a good idea to create idle callbacks for each item.  When users have
thousands of items or more, things can really slowdown.

Don't schedule an update if we fail to download an icon.  We'll try on restart
anyways, so it's not a big deal.  Scheduling the update has the same
potential problems as above, we can end up scheduling thousands of timeouts
under certain circumstances.