Performance improvements for updating watched directories.
Lower the amount of time we allow for the bulk update code to find items.
Spending 0.5 seconds adding items seems like too much, because that doesn't
take into account the time needed to commit the changes and update our view
trackers.

Tweaked the code that updates view trackers during bulk updates.  The old code
was based on the idea that a lot of objects would be changes, so it would be
faster to calculate all the ids for a table that match a view.  However, if
not a lot of objects have changed, we should just use our normal strategy of
checking only the rows that we know have changed.

This hopefully will fix the directory update part of #17785.