Added grouping functionality to InfoList.
This is for hybrid mode.  This will allow us to know the position of a single
cell inside the group of cells for an album.

Implemented infolistplat_iter_for_node() on OS X.  It's easy to do now, and I
wanted to use row_for_id() in the unittests.  row_for_id() depends on
infolistplat_iter_for_node().