Fixes for converting paths to URLs
Fixed directory_feed_url_to_path(). I'm not sure what I was thinking when I
first wrote it, it was just totally wrong.

Removed code to try to handle invalid unicode values in
path_to_directory_feed_url().  We were still highly likely to fail, so it
makes more sense to get a crash report closer to the source of the error.

Fixed resources.url() on linux and OSX.  Both of them assumed that paths were
utf-8 encoded.

Started using urllib.pathname2url(), it seems nicer to use a function that's
made for what we are doing, rather than using urllib.quote() and modifying the
safe chars.

Added unittests for the bugs.