Created new view for the hybrid album renderer
Called this view "album view", since I think that's closer to how we'll
advertise it to users.  Created a new, ugly, placeholder image for that and
put it in the middle of list and standard view.

Made album view share the same state with list view for which columns are
enabled and what size they are.  This is probably not the best way, but it's
simpler to implement for now.

Removed code that lays out the left/right buttons on the view switcher.
Instead, I just made all the images the same size.  I think this is a lot
nicer programically and also for the designers.

Removed DeviceItem.build_header_toolbar(), which was overriding the parent
method, but not changing any behavior.

Added a bunch of widget state stuff to support the new view