Refactored item filters
Store the active filters using strings rather than integers.  This makes it
easier for extensions to add them because we don't have to keep a global table
of bit flags.

Added itemfilters module.  It contains classes that handle the logic for
filters.  It helps split up the WidgetStateStore.toggle_filter() and
ItemList._should_show_item() methods, both of theme were really long.

Removed new_only ItemList and a bunch of unused ItemList subclasses.  If we
need these again we should write an ItemFilter to handle it.

Simplified code to add filters to ItemListTitlebars and removed confusing
system of mixin classes to setup filters.

Removed the ItemListTitlebar.toggle_filter code.  I couldn't understand what
was happening there, but I think a lot of it was unused.  Replaced it