Added support for extensions to add item filters.
Added widgetsapi module.  It's like api, but for widgets-specific things.

Added new interfaces called ExtensionItemFilter.  Right now it's exactly the
same as ItemFilter, but I figure it's better to separate them so that we can
change ItemFilter and keep the ExtensionItemFilter API frozen if we want to.

Added hook function capability to ExtensionManager.  This allows you to call a
hook function for all loaded extensions that support it.  There's also some
code to collect the results of the function calls.

Add the path for the extensions directory to sys.path regardless of if we find
an extension there.  I don't think this makes much difference either way and
it's needed to load extension hooks.

Implemented example item filter extension.