Added storage capabilities to the extension API
Created the StorageManager class that extensions can use to store stuff.
Created two systems for storage, one for basic uses, and one for advanced
usess.  Both use an SQLite database file to store things.

Added an ExtensionContext argument to the load() function for extensions that
stores objects specific to that extension.

Used extension module name to create the database file.  That must be unique
and seems pretty natural as a filename.

Made all unittests use a sandboxed support directory, instead of just the fast
resume test.

Didn't add support for migrations.  If extension writers want this, we can add
it later.