Fixed threading issues on windows
We removed the threads_enter() and threads_leave() calls, which causes the GTK
thread not to run on my machine.  Instead of adding them back, switch to using
gobject.threads_init().  According to the PyGTK FAQ, this is what we want to
use when we have multiple threads, but only one of them acesses GTK.