Make gstreamer work better on windows.
This changeset uses gdk_window_ensure_native() to give us a native HWND to
pass to gstreamer.  Before we were passing a client-side window to gstreamer
which confused it and made it take up the whole window area.

Events don't seem to be working correctly.  Both double-click and motion
events aren't firing on windows.

I also added a pygtkhacks function to calculate where inside a native window a
client-side window is, but it's not used yet.  I think it will be useful for
making XULRunner work again, and we may even end up using it for gstreamer as
well.