bz17792: Replace failed_soft() with logging.error() in case of item having gone MIA in forget_object().
A database item when being deleted can go missing for various reasons.
Most common case is probably a variant of an item being tracked by
Miro being moved, renamed or deleted behind Miro's back.  This will likely
cause cascading failures when #17917 is fixed. (e.g. icon cache not
available when item is deleted).

For the particular traceback detailed in #17792, it seems to be a case
of the main torrent along with its children being selected for deletion,
and the main torrent was deleted first, which seems to remove the children
as a side effect, so when we get around to removing the children it's no
longer there.