Added album view mode for the videos tab.  (#18186)
For the videos tab we use either the user-set show name, or the feed name for
the album.  For the image, we try to use the feed image, but usually just fall
back to a default.  I'm not sure what the best thing to use here is.

Refactored MultiRowAlbumRenderer a bit.  Removed logic to pick which data to
for which mode and moved it to separate subclasses.