bz17973: Avoid IndexError on list pop.
We should check to see whether a list is empty before we call pop() if we
don't intend on catching IndexError.

I don't exactly know why this happens, maybe it's some bad edge case with
playing external, which I know screws around with stuff to make things work
which may break invariants which should not necessarily be safely assumed,
and/or something to do with running callbacks.