Use GPL version of OSSBuild gstreamer 10.7 beta 4
This works in conjunction with a miro-binary-kit-win change to fix 2 issues
with audio on windows

The main issue was that AAC audio was not supported.  Using the GPL builds
adds the faad element, which decodes AAC.

The other issue was that the dsoundsink would only work with 44100 khz audio.
I'm not sure what the problem was, but changing to the GPL version seems to
have fixed it.

Change the code to pick our audio/video sinks to be platform-specific.  On
windows, we know which sinks we want to use, so we don't need to have all the
testing code that linux uses.  On linux, we just use the old code from
portable.

Changed how we construct AudioRenderer and VideoRenderer.  Instead of having
the renderer module define the clases, it defines a factory method that
creates the objects we need.  This is more flexible and it was the simplest
way to have different sink picking code on different platforms.