bz18239: Don't trust the HTTP HEAD response from the server too much.
We use the HTTP HEAD request to get a bit of info about the item before
proceeding to download but it doesn't always work.  Either the server
may not support it, or maybe it it returns some other status code
when the item's really there.

So, if it returned a response code that we did not expect and it was
a HEAD request, try to download anyway and let the download code deal
with failure.