bz:18264: In httpclient, don't scribble filesystem on HTTP error.
In bz:18239, we changed the behavior of httpclient so that it does not
rely on HEAD, and tries to grab the file anyway.  Unfortunately, after
thinking more, that is prone to breakage too.  The problem is the
pycurl.WRITEDATA: it scribbles stuff regardless of what body you get,
so we can get data on successful fetch (good), or non-successful data (bad).
In the latter case it scribbles over the filesystem unnecessarily and
makes the system think you have downloaded a file when in fact the file
is bogus.  Before, this symptom was more rare since it can only happen
on the HEAD hack for specific HTTP error codes, now it can happen more often
since we try to do a fetch anyway regardless of what the HEAD said.

The fix is to use WRITEFUNCTION instead of WRITEDATA, and ensure we only
perform a write on data where the http code is recognized to be clean.