bz18104.  Catch ValueError in next_free_{filename,directory}().
The next_free_{filename,directory}() functions can throw ValueError if
a set number tries exceed a threshold.  But, code using this does not
catch ValueError, leading to uncaught exception.  Worse, wrap these
around a @returns_filename decorator, which catches ValueError and
reraises ValueError with a different message, so the user/developer
is left confused as to why there is an apparent internal failure with
this suite of functions.