Fixed subtitle font selection on windows (#15227)
Now that we use gstreamer, font selection is way simpler and works correctly
with unicode.  We just use pango to get the list of font families to select.

Moved renderer code to select fonts to portable.  On liunx, there's no UI to
change SUBTITLE_FONT, but it doesn't hurt to have it there.  Users could use
gconf-editor to change it if they really wanted to.