bz17155: Address part (b) of bz17155.
Items jump around when you click on pause all or resume all or cancel all.
The problem is the commands to the remote downloader for each individual
item is sent individually and the response is also replied to individually
meaning that when the items update they are resorted, so for n
items you have n updates and thus n resorts, making the UI jerky.

It'd be nice to batch all of this stuff up so that it's one single ipc
but it's too hard to fix, so when we use a pause all, resume all or
cancel all, first, disable the resort mechanism on the downloads
controller, then, perform an ipc rendezvous with the downloader when
all the status updates have been sent so we can re-enable the sort.