Refactored how we setup the sorts menu.
Moved the code to the Display class.  I think it makes more sense to have it
there, since the display is more intimate the data we need than the menu code.

This will be even more true once we make list view and album view store
separate sort data.  Once that happens we need to know which view is selected
to know which columns are enabled.  The menu code doesn't have a clean way of
knowing that.

Made code to toggle which columns are enabled use a similar system.