Made album view column resizable (#18368)
Moved code that add "multi-row-album" to the list of columns from
itemlistwidgets to widgetstatestore.  This makes that column work more like
the others.

Moved code to set up manditory columns to its own method.  We need this
because we want to put name on the end, but album at the start.