Started overhauling the menu system
Replaced old system with something more direct.  The platforms define MenuBar,
Menu, and MenuItem and the portable code uses them to create menus.  This
should help us move more code to portable.

The benefit is it makes it much easier to manipulate the menus from portable
code.  We also get to delete a ton of GTK code that wrestles with UIManager to
do what we want.

Some of the old code is now commented out since it needs to be rewritten for
the new system.  The main things to do are: enabling/disabling groups,
changing labels, and dynamically updating the audio track menu and the
subtitle track menu.

Moved GTK keymap code from window.py to keymap.py.  Moved portable key
handling code to its own module, so that it can be imported by both portable
and platform code without any hassles.  Changed the way we handle the "MOD"
key.  Instead of setting it to either CTRL or CMD, we just give it its own
value and let the platforms map it to either control or command.

The new code only supports one shortcut per menu item on both GTK and OSX.
Before we supported multiple ones on GTK, but it took a lot of crazy code to
do so.  We can achieve the same result using key-press handlers anyways.