Reimplemented enabled_groups and state_labels
I kept most of the old code, but moved the code that manipulates the MenuItems
to portable.  I'd love to get rid of the code entirely, but that's another
project.  At least it's all in portable, so it should be easier to fix in the
future.

Reimplemented platform-specific code that relied on MenuItem groups since I
didn't want to have the platform code import the portable code.

Added a method in itemlistcontrollermanager to quickly get information about
the selected items without having to iterate through them.  This makes it
easier to split up code that enables/disables menu items.

Renamed MenuStateManagar to MenuManager.  It now actually manages the menu
items rather than just updating a bunch of state variables.  Removed the
enabled-changed signal.

Totally disabled the sort menu for now since we only support groups on actual
MenuItems, not CheckMenuItems.

Fixed bug in new OSX code to enable/disable menu items.