Reimplemented view menu logic
Mostly this is just removing a bunch of old stale code.  One nice thing is
that we now hide the sorts menu on GTK as well as OSX.

Added some generic code for dynamic menus.  I think it should work for the
audio track and subtitles menu too.

Made RadioMenuItem a subclass of CheckMenuItem and gave both of them a
set_state() method.  Added show/hide methods to Menu.