Reimplemented changing audio tracks
Moved the audio track code to the Player class and then to the PlaybackManager
class. This makes it accesible from portable code.  Used the gstreamer API for
this and changed the quicktime code to match it.  Made a base class for
AudioPlayer and VideoPlayer on GTK.

Fixed the RadioMenuItem constructor on GTK.  Added get_label() and get_state()
methods to menu items.  Made get_children() return a new list to make it
easier to remove the children.

Refactored the MenuUpdater code a bit.