bz18446: Defer loading of right hand side displays.
Defer the loading of right hand side displays because this is a relatively
expensive process that outstrips the time it takes for keyboard navigation
to take place.  If the display was only transient and will be destroyed
again because the navigation has not yet finished, then just not bother
creating it.