Added some optimizations for menu changing code
Added code to skip calculations to update the menu code if we know that we
won't change it.  Since update_menus() is called fairly often, it seems pretty
bad to always run all the menu update code.

Refactored who calls update_menus() a bit.  Made ItemListController do it
when we change our item infos instead of WidgetsMessageHandler.  I think
that's a more natural class to handle that.