Don't leave files in the echoprint directory
It's annoying because echoprint is a submodule, so it means that that
subrepository is always dirty.  We don't have the ability to add something to
echoprint's .gitignore, so let's move the file to our build directory.

Also cleaned up the setup.py script a touch.  I think this way is much simpler
to use.