bz18510: Change subprocess module innards for proper termination.
Make 2 changes.  Firstly, we run the the message loop dispatch in the
main thread and farm off the handler to a separate thread whereas
previously the job role was of the threads was reversed.

Secondly, don't bother checking the returncode attribute on quit.
If it failed to kill the subprocess because of whatever reason
then just ignore it.