Revert part of f9f0462.
Revert the part of f9f0462 where it switches the role of the pipe thread
and the handler thread.  With this revert, it reverts to running
the pipe thread in a secondary thread and the handler in the main
thread.  The reason we need to do this is for QuickTime where the QuickTime
objects must be created on the main thread.

This reintroduces the problem where the quit message (denoted by
msg = None) is not immediately received and is backlogged.  But this is
less of a problem because if we didn't quit the main instance will force
us to quit with a kill() anyway.

All of this is volatile and may change soon anyway so I'm not going
to bother thinking too much about getting it right.