Added a thread pool to the worker thread
The main goals for this were:
  - Feedparser gets priority over mutagen and movie data
  - If we have both mutagen and movie data calls, we try to schedule them both
    together, since one is disk-bound and one is cpu-bound.

Implemented it with a priority queue with round-robin scheduling for tasks
with the same priority.  Hopefully this will be good enough.

Also added a queue on the metadata manager side to avoid flodding the worker
process with too many mutagen/movie data requests.

Added code in unittest setup to reset the worker task queue in tearDown().
Otherwise if one test leaves something in there it will mess up the rest of
the tests.  Added more printouts when we get a failed_soft() call in the unit
tests.

Added message for the worker process to signal when its ready.  I think this
is good in general, but it also makes the unittests more deterministic.

Changed subprocessmanager._send_subprocess_error_for_exception() to be a
public function and started using it in workerprocess.