Run movie data tasks on the main thread in workerprocess.
This is because of limitations in both gstreamer and quicktime:

We need to run them all on the same thread or our current gstreamer code will
break.

Because of limitations in QTKit before OSX 10.5, the thread must be the main
thread.

Removed the worker system test.  Those tests were not very deterministic,
since they depended on the order of how things got processed in the worker
process.  Now it's even harder with this hack for movie data.  I'd like to
test this stuff, but I don't have the time to write a good one.