bz18549: Fix crash with tablist iter abuse.
We must be very careful when dealing with iters as using a stale one
can cause the entire program to crash.

Here, we make sure that when we forget the child iters we really forget
all children, and also that we remove them from the view.  As well, we
make sure that we handle no tabs selected last, after all changes have
been committed to the table model, so that the tablist iter
view is consistent.