Changed how we store cover art
Made it so that we use the album name to store the cover art path.  This makes
it easy to have items share the same cover art if they need to and to avoid
writing out the cover art data for a single album more than once.

Changed how cover art gets created.  Now the caller of CoverArt.write_file()
just passes in a path to write to.

Removed the CoverArt.from_file() method.  I'm not sure how that should work in
the new system.  I don't think it was being called anyway.

Removed unittest for upgrade 166.  It's not very easy to maintain as the
metadata schema keeps changing.  Also that upgrade method shouldn't change, so
having a unittest for it dosen't make a ton of sense.