Started echonest work.
Added some code to handle echonest metadata, but only hooked up the first half
of it which is sending the file to ENMFP.  Used ENMFP instead of echoprint
because the song database is just so much bigger.

Bumped the number of threads to 4 since we will often be running enmfp in a
thread.

Refactored the metadata tests by moving all the code that makes a path from a
filename into 1 method.