Added fine-grain control over echonest/internet lookups
Removed the "ECHONEST_ENABLED" pref and replaced it with a new system.  The
new sytem is cleaner and lets users run internet lookups for individual files
and as well as not use internet data.

The new system is:
   MetadataStatus have an net_lookup_enabled flag.
   If it's True we will:
      Put the file through _EchonestProcessor
      Use the data from echonest / other internet sources
   If it's false we won't do either.
   The NET_LOOKUP_BY_DEFAULT controls its initial value, and there's an api to
   change it for specific files

Generally tried to replace "echonest" with "net lookup" since we are really
querying a variety of sources, and those sources may change from version to
version.

Made the net_lookup attribute available from ItemInfo, though we're not using
it yet.