Allow DDBObject to use other databases.
Made it input a parameter that includes the DB, the ViewTrackerManager, etc
and use that rather than the global variables.  By default will still use the
global ones, so this isn't a functional change.  But it opens the door for
sqlite databases on a device.

Altered the unittests to handle this.