Added MetadataManager for devices.
Made it so each device gets a metadata manager, connected to a LiveStorage
object on its database.

Added a function to upgrade old device databases to the current system.

Added an attribute to LiveStorage and DeviceDatabase that specifies if they
created a new database or not.

Fixed a bug with the regular items where we wouldn't remove metadata for
removed items.

Still to figure out:
   - When should we restart incomplete metadata on devices?
   - What should we do if we run into an sqlite error when writing to a device
     db?