Better error handling for devices.
Added code to deal with save errors for device databases.  Added code to
simulate getting a save error for a database.  Refactored the cleanup code for
this case.  I did it right, it will work exactly the same for the main
database.

Changed integrity check code.  Now we only run it if we get an error running
an sqlite statement.  The value is used for the device error handler, but not
yet for the the error handler for the main device.

Made MetadataManager deal better with not being able to write to its
directories for cover art/screenshots.

Use mount point rather than device point for logging errors.  mount is always
set while device_name may not be.

Preallocate space on the device so that we avoid having to run any of this
code.