Fixed some issues with ProgressCountTracker
Updated the docstring to specify which parts of the API are the same and which
are different.

Fixed a couple calls from LibraryProgressCountTracker to ProgressCountTracker
where the library count tracker was passing in file_type instead of metadata.
ProgressCountTracker just ignores the argument, but we should still pass in
the correct data.