bz18664: Make sure that the sharing manager is available before using.
Just a few days ago we put in some code that delayed the start of
the sharing manager and tracker to spread the workout out a little
bit.  Sadly, these two things are used at shutdown to checking if
there are outstanding connections (easy to fix) but also used
to manipulate sharing preferences (harder to address).

Let's put the startup of the sharing subsystem back where it was, since
that should be pretty quick, but track the items a bit later.