Better subprocess management
Try to restart the child process if it quits before we tell it to.

Added a cooldown period before restarting subprocesses.  We won't start more
than 1 per minute.

Better logging when we restart subprocesses.  In particular, log
GetLastError() on windows when we fail.

Most of these are based on trying to fix #17804