Fixes for movie data hanging the worker process.
Added a check in the main process to terminate the subprocess if it's been
working on a movie data task for too long.

Added a check on the windows MDP script to cancel its operation if it notices
that it's been taking too long.

Fixed thread initialization code on windows MDP.  The old way was making the
process just hang for me.