bz18541: Don't store displaytext output in ItemInfo
The issue was that we would create an ItemInfo with a translated string, then
the user would switch languages and we wouldn't keep up.

Calculating the display_* attributes was for performance issues, but now we're
calling a python method anyways to get the value, so I don't think it's needed
anymore.  I profiled a bit and didn't see any difference.