bz18277: fix context menu for the album view column
We do this by selecting all items in the current album group.  I couldn't
think of a good way to do this in portable, so we have to do it in
platform-land.

Skip the hotspot code on GTK when the right mouse button is clicked.  This is
what we do on OSX and I think it makes more sense.