bz18691: Don't call clamp_text() on options when creating option widget.
clamp_text() cuts text off so you could end up with 2 different entries
that become the same when clamp_text() is used.  Not only will indexing
into the original list be wrong when you retrieve the index from the option
widget but some options will not appear (because they are the same).