bz18756 part 2: fixes for devicedatabaseupgrade
Check for items that aren't using the metadata system each time we open a
device instead of only when we create the SQLite database.  This works better
when users go back and forth between versions.

We set some values for mdp_state, title_tag, and metadata_version, for
DeviceItems so that they're present for older miro versions.  The values we
set ensure that old versions won't run mutagen or MDP on the items, but should
see the metadata when never versions run mutagen/MDP.