bz18756: Device importing improvements
Refactored the code that deals with failures when converting old databases.
Now it handles it on a row-by-row basis.  If one conversion fails, we insert
values as if the file was just added to the system and continue with the rest
of the items.  Changed _do_import_old_items() into a function object to try to
make it at least semi-understandable.

Added code to deal with databases with older versions by resetting them.  This
is a fairy heavy handed approach, but it's simple and only affects users who
have been running the nightlies.

Changed LiveStorage._get_version() to be a public method.  Made
reset_database() optionally initialize the schema as well.