Fixes based on willkg's comments.
Use variables for constants rather than just setting the values.  I realized
that devicedatabaseupgrade is not like databaseupgrade in that we can safely
import other modules in that code and actually use the value.

Fix for our code that resets the database if the version is different than
what we expect.