bz18752: improvements to the net lookup pref panel
Track the number of items that we're running internet lookups and display that
in the pref panel.

Disable the net lookup buttons until the backend finishes or if they won't do
anything.

This replaces the old code that popped up a dialog when the backend finished
handling the remove all lookup data button.

Added a utility method to call a function using add_idle(), but avoid
scheduling it twice.

Changed the pref panel code slightly so that we can easily access the panel
builder objects.