bz18632: Fix shutdown on Windows.
Windows gets very sad if we don't get a chance to free the C-based
embedding widget; and relying on the garbage collection to call destroy()
isn't good because you could still have open references.  But we don't
want to call destroy() explicitly to destroy objects in case there is still
someone using it.

But empirically it seems changing it to just destroy the embedding window
is not enough so there must be something else involved.  Anyway, revert
that part of the change for now so that it all works.