bz13842: re-layout things when we hide/show a scroller.
The code that I just changed was added to fix #13842.  There's a lot of cases
there and most of them worked with or without the code.  The one case that
didn't seem to work for me was when you made the tablist grow by expanding a
section.  This change fixes that issue.