include clipping on the right side (OS X)
b83ac03a introduced a bug where the clipping rectangle no longer included the
space on the right side if the columns didn't cover that space.  This happened
for me during a column resize.  The change is to include the space to the right
if it isn't currently part of the right-most column.