bz18839: prevent nested size-allocated signals.
tile() can be called in response to size-allocated.  In the ticket report,
what happened was that size-allocated was emitted when we showed a table view,
then we set the initial widths of the columns, which called tile(), and tile()
called viewport_repositioned() which started a chain of calls that ended up
with another size-allocated signal.

Instead of calling viewport_repositioned(), call invalidate_size_request()
which will do the same thing, but it defers action a bit.  This is especially
good in that situation, because we would run that chain of calls for each
column, instead of once at the end.