bz18870: alternate fix for 17823
Looks like the root cause was the changeset 33df4b57.  That one made it so
that we checked the selection when we got the structure-will-change signal.

TreeTableModel.remove() had a bug where it would:
 - emit structure-will-change
 - forget the iter
 - call the base class remove() which will
   - emit structure-will-change again
   - remove the node from the list

_check_selection() was depending on that behavior because it would catch the
WidgetActionError during the second structure-will-change signal.

Things could crash when we called remove() twice.  In that case, we haven't
called TableView.reloadData() yet, so the first row removed could still be
selected.  This time when we emit structure-will-change the node will actually
be removed, and this is what causes the crash.

This changeset:
  - Fixes TreeTableModel.remove()
  - Adds an alternate fix for #17823