bz18886. invalidate cached thumbnails
YouTube video names are pretty common (3.*.jpg), so when we weren't clearing
the cache when changing the underlying thumbnail, the frontend was getting
confused.  We do this by attaching a function to those thumbnails which checks
whether the thumbnail has been updated.  It's possible that this will cause
slowdowns, but it should only affect visible thumbnails and there aren't many
of those at any one point.