Draw background behind everything for the album/artist column.
While testing the bug, I noticed that on GTK you could see the selection box
behind the VODO image.  I'm pretty sure this was because there was
transparency in the image file.  To fix this, fill the entire area with our
background, instead of just the parts surrounding the image.