bz18806: speedup fetching MetadataStatus objects in set_net_lookup_enabled()
Before we were fetching each one individually by path, even if we were
changing the value globally.  This speeds up that case by making a view that
contains all the MetadataStatus objects that would need to be changed.  This
is most noticable when MetadataStatus objects haven't been loaded yet and put
into the cache.